API/requirements.txt
python-dotenv
Flask
spotipy
pandas
flask_cors